Require config for OpenAPI spec generator.
