+++ example/index.ts
@@ -1,16 +1,5 @@
-import { createConfig, createServer } from "../src";
+import { createServer } from "../src";
+import { config } from "./config";
 import { routing } from "./routing";
-
-const config = createConfig({
-  server: {
-    listen: 8090,
-    upload: true,
-  },
-  cors: true,
-  logger: {
-    level: "debug",
-    color: true,
-  },
-});
 
 createServer(config, routing);
